Move Workspace implementation to AdapterWorkspace
- Move Adapter<C> interface to meta/pouch/types
- Add meta/pouch/workspace module with AdapterWorkspace<C> class, which
  accepts an Adapter<C> and implements Workspace<C>
