+++ packages/db/src/meta/pouch/index.ts
@@ -1,14 +1,17 @@
 import { logger } from "@truffle/db/logger";
 const debug = logger("db:meta:pouch");
 
-export { Definition, Definitions } from "./types";
+import type { Collections } from "@truffle/db/meta/collections";
+import type { Workspace } from "@truffle/db/meta/data";
 
 import * as Adapters from "./adapters";
 export { Adapters };
 
-import type { Collections } from "@truffle/db/meta/collections";
-import type { Workspace } from "@truffle/db/meta/data";
-import type { Definitions } from "./types";
+import { AdapterWorkspace } from "./workspace";
+
+import type { Adapter, Definition, Definitions } from "./types";
+export { Adapter, Definition, Definitions };
+
 
 export const forDefinitions = <C extends Collections>(
   definitions: Definitions<C>
@@ -19,7 +22,12 @@ export const forDefinitions = <C extends Collections>(
 
   debug("Initializing workspace...");
   // @ts-ignore
-  const workspace = new constructor({ definitions, settings });
+  const adapter: Adapter<C> = new constructor({ definitions, settings });
+
+  const workspace = new AdapterWorkspace<C>({
+    adapter,
+    definitions
+  });
 
   return workspace;
 };
